refactor(pa-cockpit): replace the 14-export surface with one PaSectionsRouter
Review round on Task 9: packages/frontend's PASectionRouter.tsx and
packages/pa-demo's DemoSectionRouter.tsx carried byte-identical
MONITORING_IDS/VOORTGANG_IDS/beheer-panel dispatch and an identical
db-overzicht/db-nieuw key-remount branch. That is section-id grammar about
the package's own ids, not host knowledge, so exporting fourteen
components and asking every host to hand-maintain it would have kept the
vendored fork's most-duplicated behaviour, just formalised.

- New packages/pa-cockpit/src/components/PADashboardV2/PaSectionsRouter.tsx:
  the package-owned dispatcher (vandaag/agenda/feiten, monitoring tabs,
  voortgang views, db-overzicht/db-nieuw, the five package 'beheer' panels,
  the dossier fallthrough and the deleted-dossier placeholder), moved
  verbatim from the old frontend router. Its own file header documents a
  composition-order requirement the review's sketch did not account for: a
  host must check its own exclusive ids FIRST and use PaSectionsRouter as
  the unconditional tail, not the first attempt -- its placeholder branch
  is unconditional, so calling it before a host's own switch would swallow
  host ids like 'profiel' into "dossier not found" instead of ever
  reaching them.
- packages/pa-cockpit/src/index.ts now exports exactly one new value,
  PaSectionsRouter, replacing the fourteen individual component/type
  exports added earlier. Vandaag, Issuekaart, Monitoring, AgendaView,
  Voortgang, FeitenView, MONITORING_TABS, the five beheer panels,
  Dossierbeheer, usePaData, Prioritering, KompasViz, MonitoringTabId and
  VoortgangView are all back off the public surface.
- New src/index.test.ts pins the package's exported value names so the
  next widening is a visible, deliberate diff.
- packages/frontend's PASectionRouter.tsx shrinks to a 7-case switch over
  its own exclusive ids (profiel, rollen, the four IOU sections,
  gereedschap-overzicht), delegating everything else to PaSectionsRouter,
  and now takes the package's own PaSectionRouterProps type instead of a
  hand-rolled duplicate.
- Test coverage relocated with its subject, not lost: the assertions that
  tested package-owned dispatch moved into PaSectionsRouter.test.tsx
  alongside the code they test; frontend's test file shrinks to mocking
  one export instead of twelve. Net test count across both workspaces grew
  (1187 -> 1197).
- Restored PaSectionRouterProps' dropped JSDoc and moved the "Getters, not
  snapshots" comment to sit above the getters it explains.

See task-9-report.md's addendum for the full verbatim-duplication diff, the
composition-order reasoning, and what packages/pa-demo (Task 11, untouched
here) will be able to delete once it delegates to PaSectionsRouter too.
